llm_service/requirements.txt
pyyaml>=6.0
faiss-cpu>=1.7.4
numpy>=1.24.0
requests>=2.31.0
fastapi>=0.104.0
uvicorn>=0.24.0
pydantic>=2.0.0
sentence-transformers>=2.2.0
jinja2>=3.1.0

